Refactor workers to employees
Cambiso en:

- workers_controllers.js: Todo el contenido de este archivo se pasó a employee_controller.js.

- index.js: Se eliminó la ruta de /workers.

- workers.js: Se eliminó el archivo y las rutas se pasaron a employee.js.

- router.js: Se eliminó la importación de las rutas /workers.

- existing_inventory.ejs: Se movió el archivo de la carpeta /workers a /employees.

- _navbar.ejs: Vínculo a la vista de inventario para trabajadores.
