+++ src/index.js
@@ -54,8 +54,7 @@ app.use((req, res, next) => {
 const router = require('./routes/router.js'); 
 app.use('/', router.authentication); 
 app.use('/', router.general); 
-app.use('/admin', router.admin);
-app.use('/workers', router.workers);  
+app.use('/admin', router.admin); 
 app.use('/employee', router.employee)
 app.use('/cart', router.cart);
