+++ src/routes/router.js
@@ -6,6 +6,5 @@ router.general = require('./general_routes');
 router.admin = require('./admin'); 
 router.employee = require('./employee');
 router.cart = require('./cart_api'); 
-router.workers = require('./workers');
 
 module.exports = router;
