+++ src/views/partials/_navbar.ejs
@@ -92,7 +92,7 @@
                 >
             </li>
             <li class="navigation__item">
-                <a href="#" class="navigation__link"
+                <a href="/employee/inventory" class="navigation__link"
                     ><div class="navigation__icon"><span class="icon"><img src="/icons/th-list-solid.svg" alt="Ícono inventario"></span></div>
                     Consulta de inventario</a
                 >
